+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## Unreleased
 
+* Handle URL encoded links
+
 ## v0.1.4 (October 31, 2022)
 
 * Return an error code if a site fails the test